Renaming the dynamics portion to legion, since it's borrowing heavily from that project in terms of concepts and otherwise anyway.
