+++ src/legion/atom.rs
@@ -1,8 +1,8 @@
 use std::collections::HashMap;
 use std::ops::Sub;
 use uuid::Uuid;
-use crate::dynamics::ff::ForceField;
-use crate::dynamics::ff::Elements;
+use crate::legion::ff::ForceField;
+use crate::legion::ff::Elements;
 use super::particle::HasCharge;
 use super::particle::HasMass;
 use super::particle::HasPhysics;
